#81 fix bug + fix tab
