+++ src/pages/student/userProfile/bookingHistory.tsx
@@ -1,0 +1,10 @@
+
+function BookingHistory() {
+  return (
+    <div>
+      <div></div>
+    </div>
+  )
+}
+
+export default BookingHistory
