+++ src/pages/student/userProfile/programHistory.tsx
@@ -1,0 +1,16 @@
+import "./userProfile.scss";
+
+function ProgramHistory() {
+  return (
+    <div className="user-profile__tab-content">
+      <h3>Lịch sử tham gia chương trình</h3>
+      <ul>
+        <li>Chương trình A - 01/01/2023</li>
+        <li>Chương trình B - 15/02/2023</li>
+        <li>Chương trình C - 20/03/2023</li>
+      </ul>
+    </div>
+  );
+}
+
+export default ProgramHistory;
